refactor: integrate `inline` into `commands`
